requirements.txt
accelerate==1.0.1
diffusers==0.32.2
einops==0.8.1
flash_attn==2.5.6
huggingface_hub==0.25.0
imageio==2.34.0
lmdb==1.5.1
mc==0.0
modules==1.0.0
numpy==1.24.4
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
pandas==2.0.3
Pillow==11.1.0
pyiqa==0.1.12
pytz==2021.3
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
skimage==0.0
tensorboard==2.14.0
tensorflow_io==0.37.1
tqdm==4.66.1
transformers==4.37.2
typed_argument_parser==1.10.1
wandb==0.19.8
xformers==0.0.24